refactor: change the estructure project
